+++ src/figdraw/fignodes.nim
@@ -4,11 +4,6 @@ export tables, hashes
 import ./figbasics
 export figbasics
 
-when defined(figdrawNativeDynlib):
-  {.pragma: nativeAbi, exportabi.}
-else:
-  {.pragma: nativeAbi.}
-
 type
   DrawableKind* = enum
     dkLine
@@ -230,8 +225,6 @@ proc remappedNodes(list: RenderList, insertIdx: int, parentIdx: FigIdx): seq[Fig
 proc relevelNodes(nodes: var seq[Fig], lvl: ZLevel) =
   for node in nodes.mitems:
     node.zlevel = lvl
-
-{.push nativeAbi.}
 
 proc drawableLine*(a, b: Vec2): DrawableOp =
   DrawableOp(kind: dkLine, a: a, b: b)
@@ -553,5 +546,3 @@ proc addChildren*(
 
 proc contains*(r: Renders, lvl: ZLevel): bool =
   r.layers.contains(lvl)
-
-{.pop.}