+++ tests/tsiwin_redraw.nim
@@ -3,7 +3,7 @@ when defined(linux) or defined(bsd):
   import std/os
 
 when defined(useNativeDynlib):
-  import figdraw/dynlib
+  import figdraw
   from figdraw_native_abi import nil
 else:
   import figdraw