+++ tests/tsiwin_resize_presentation.nim
@@ -1,6 +1,9 @@
 import std/unittest
 
-import figdraw
+import chroma
+import figdraw/commons
+import figdraw/fignodes
+import figdraw/figrender
 import figdraw/windowing/siwinshim
 
 when defined(macosx) and (UseMetalBackend or UseVulkanBackend):